connection/requirements.txt
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
pyserial==3.5
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
